minor bootstrap stylin, moved logout button to right, button styling and members table styling
