+++ predictions/client/src/components/userdashboard/DashboardRight.jsx
@@ -1,10 +1,10 @@
 import {useState, useEffect, useContext} from 'react'
 import axios from 'axios'
-import {useParams, Link} from 'react-router-dom'
+import {useParams, Link, useNavigate} from 'react-router-dom'
 import UserContext from '../../context/userContext';
 
 const DashboardRight = () => {
-
+  const navigate=useNavigate()
   const { user, setUser, leagues, setLeagues } = useContext(UserContext)
   const [notUsersLeagues, setNotUsersLeagues] = useState([])
 
@@ -23,6 +23,12 @@ const DashboardRight = () => {
 }, []
 )
 
+    // Logout user
+    const handleLogout = () => {
+      setUser(null);
+      document.cookie = 'userToken=;';
+      navigate('/login');
+    };
 console.log("LEAGUES", leagues)
 // const leaguesToJoin=leagues.filter(league=>league.user!=user._id)
 
@@ -38,6 +44,8 @@ console.log("LEAGUES", leagues)
             
             </div>
       ))}
+      {/* Logout button */}
+      <button onClick={handleLogout} className="btn btn-danger">Logout</button>
     </div>
